mist/requirements.txt
torch==1.11.0
transformers==4.28.0
numpy==1.22.0
punctuator==0.9.6
tqdm==4.42.1
ffmpeg_python==0.2.0
pandas==1.1.1
ffmpeg==1.4
pipelines
scikit_learn==0.24.2
CherryPy==18.6.0
h5py==3.8.0
matplotlib==3.5.3
decord==0.6.0
opencv-python==4.7.0.72
Pillow==9.4.0
